Change title to include "My game development journey"
The old title "Increasing my contribution to Zig to $200/mo" was a bit
confusing/misleading (e.g. [this](https://news.ycombinator.com/item?id=26788644))
and I suspect many dismissed the article based on the title alone ("feels click-baity")

I also received some feedback privately that the title felt like a bit of bait-and-switch.

My hope is the new title:

> My game development journey & why I'm increasing my contribution to Zig to $200/mo

Will more clearly let people find it relevant (or not!) if they so choose.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>