add Plausible: privacy focused analytics
I hate analytics in general and think they are dumb, but it would
be cool to know how many people are actually finding my writing
interesting. I would never add Google Analytics or anything like
that, but Plausible.io seems like a nice middle-ground:

> By using Plausible Analytics, all the site measurement is carried
> out absolutely anonymously. Cookies are not set and no personal
> data is collected. All data is in aggregate only. The website owner
> gets some actionable data to help them learn and improve, while the
> visitor keeps having a nice and enjoyable experience.

Additionally, I believe analytics should always be open and public.
If we're going to claim that analytics are to help us improve (which
I do believe they can be), then they should be public and transparent.
Otherwise, it's spying - it's that simple.

In the spirit of that, our analytics are public and can be viewed here:

https://plausible.io/hexops.com

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>